Move internal macros to separate crate, add readme and co
