+++ crates/cubecl-ir/Cargo.toml
@@ -1,9 +1,18 @@
 [package]
-edition = "2021"
+authors = [
+    "nathanielsimard <nathaniel.simard.42@gmail.com>",
+    "louisfd <louisfd94@gmail.com",
+]
+categories = ["science"]
+description = "Intermediate representation for CubeCL"
+edition.workspace = true
+keywords = []
 license.workspace = true
 name = "cubecl-ir"
 readme.workspace = true
+repository = "https://github.com/tracel-ai/cubecl/tree/main/cubecl-ir"
 version.workspace = true
+
 
 [features]
 default = ["serde"]
@@ -12,7 +21,7 @@ serde = ["dep:serde"]
 
 [dependencies]
 cubecl-common = { path = "../cubecl-common", version = "0.5" }
-cubecl-macros = { path = "../cubecl-macros", version = "0.5" }
+cubecl-macros-internal = { path = "../cubecl-macros-internal", version = "0.5" }
 
 derive_more = { workspace = true, features = ["from"] }
 float-ord = "0.3"
