+++ crates/cubecl-macros-internal/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+authors = [
+    "nathanielsimard <nathaniel.simard.42@gmail.com>",
+    "louisfd <louisfd94@gmail.com",
+]
+categories = ["science"]
+description = "Internal procedural macros for CubeCL"
+edition.workspace = true
+keywords = []
+license.workspace = true
+name = "cubecl-macros-internal"
+readme.workspace = true
+repository = "https://github.com/tracel-ai/cubecl/tree/main/cubecl-macros-internal"
+version.workspace = true
+
+[lib]
+proc-macro = true
+
+
+[dependencies]
+darling = { workspace = true }
+proc-macro2 = { workspace = true }
+quote = { workspace = true }
+syn = { workspace = true }
