+++ crates/cubecl-macros-internal/LICENSE-MIT
@@ -1,0 +1,1 @@
+../../LICENSE-MIT
