+++ crates/cubecl-macros-internal/README.md
@@ -1,0 +1,3 @@
+# CubeCL internal macros
+
+Internal procedural macros for generating additional implementations on the CubeCL IR.
