+++ crates/cubecl-macros-internal/src/generate/mod.rs
@@ -1,0 +1,2 @@
+pub mod op_args;
+pub mod operation;
