+++ crates/cubecl-macros-internal/src/generate/op_args.rs
