+++ crates/cubecl-macros/src/parse/mod.rs
@@ -9,8 +9,6 @@ pub mod desugar;
 pub mod expression;
 pub mod helpers;
 pub mod kernel;
-pub mod op_args;
-pub mod operation;
 pub mod operator;
 pub mod statement;
